requirements.txt
click
get-mac
pycryptodome